feat: run pg_regress tests after installing the last version of the extension (#1826)
We have been testing the extension by running pg_regress tests after the
migration of the extension (from the first available version to the last
one). However, we were not testing the case where the last version of
the extension is installed directly. This change adds a test for that
case.