chore: bump Wrappers version to 0.5.7 (#1945)
* chore: bump Wrappers version to 0.5.7

* add new pgrx and rust version

* feat: support 1.88 rust 0.16.1 pgrx 0.57 wrappers

* docs: update docs

* fix: reduce pressure on CI runners

Updating the dependency helping us build all our rust extensions has put
a lot of pressure on our CI runners, causing timeouts and failed builds.

We are currently working on a more permanent solution in #1745 but in
the meantime, we cap the number of parallel nix builds to 2, which
should reduce the load on our CI runners.

* fix: pg_regress test for wrappers extension

pg_regress tests always test the latest update of our extensions. 
Add the wrappers renamed types to the expected output files.

* fix: increase parallelism for build process

We have reduced the pressure on CI runner, now the build process is timing out after 3h.
By increasing the parallelism from 2 to 4, we can better utilize the available resources and speed up the build process without overwhelming the system like before.

* fix: delegate copying artifacts to nix-fast-build

Copying built artifacts to S3 should be done by nix-fast-build itself
while building instead of relying on nix-daemon's post-build hooks and
the end of the build.

* chore: bump to release

---------

Co-authored-by: Sam Rose <samuel@supabase.io>
Co-authored-by: Jean-François Roche <jfroche@pyxel.be>